Power advertise

+++ components/ble_mi_remote/ble_mi_remote.cpp
@@ -205,7 +205,7 @@ namespace esphome {
 		void BleMiRemote::powerAdvertising() {
 			NimBLEUUID* svcUUID = new NimBLEUUID();
 			NimBLEAdvertisementData* advData = new NimBLEAdvertisementData();
-			advData->setFlags(0x0101);
+			advData->setFlags((uint16_t) 0x0101);
 			advData->setManufacturerData("0100");
 			advData->setShortName("MI RC");
 			advData->setPartialServices((NimBLEUUID) "1812");
